Make "SelectionFAB" reusable
Was to tightly coupled to chapters
