+++ src/components/manga/SelectionFAB.tsx
@@ -10,37 +10,21 @@ import MoreHoriz from '@mui/icons-material/MoreHoriz';
 import { Fab, Menu, Box } from '@mui/material';
 import React, { useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import Download from '@mui/icons-material/Download';
-import Delete from '@mui/icons-material/Delete';
-import BookmarkAdd from '@mui/icons-material/BookmarkAdd';
-import BookmarkRemove from '@mui/icons-material/BookmarkRemove';
-import Done from '@mui/icons-material/Done';
-import RemoveDone from '@mui/icons-material/RemoveDone';
 import { DEFAULT_FAB_STYLE } from '@/components/util/StyledFab';
-import { SelectionFABActionItem } from '@/components/manga/SelectionFABActionItem';
-import type { IChapterWithMeta } from '@/components/manga/ChapterList';
-
-type SelectionAction = 'download' | 'delete' | 'bookmark' | 'unbookmark' | 'mark_as_read' | 'mark_as_unread';
+import { TranslationKey } from '@/typings.ts';
 
 interface SelectionFABProps {
-    selectedChapters: IChapterWithMeta[];
-    onAction: (action: SelectionAction, chapters: IChapterWithMeta[]) => void;
+    children: (handleClose: () => void) => React.ReactNode;
+    selectedItemsCount: number;
+    title: TranslationKey;
 }
 
-export const SelectionFAB: React.FC<SelectionFABProps> = (props) => {
+export const SelectionFAB: React.FC<SelectionFABProps> = ({ children, selectedItemsCount, title }) => {
     const { t } = useTranslation();
-
-    const { selectedChapters, onAction } = props;
-    const count = selectedChapters.length;
 
     const anchorEl = useRef<HTMLElement>();
     const [open, setOpen] = useState(false);
     const handleClose = () => setOpen(false);
-
-    const handleAction = (action: SelectionAction, chapters: IChapterWithMeta[]) => {
-        onAction(action, chapters);
-        handleClose();
-    };
 
     return (
         <Box
@@ -53,7 +37,7 @@ export const SelectionFAB: React.FC<SelectionFABProps> = (props) => {
             ref={anchorEl}
         >
             <Fab variant="extended" color="primary" id="selectionMenuButton" onClick={() => setOpen(true)}>
-                {`${count} ${t('chapter.title', { count })}`}
+                {`${selectedItemsCount} ${t(title, { count: selectedItemsCount })}`}
                 <MoreHoriz sx={{ ml: 1 }} />
             </Fab>
             <Menu
@@ -67,50 +51,7 @@ export const SelectionFAB: React.FC<SelectionFABProps> = (props) => {
                     'aria-labelledby': 'selectionMenuButton',
                 }}
             >
-                <SelectionFABActionItem
-                    action="download"
-                    Icon={Download}
-                    matchingItems={selectedChapters.filter(
-                        ({ chapter: c, downloadChapter: dc }) => !c.isDownloaded && dc === undefined,
-                    )}
-                    onClick={handleAction}
-                    title={t('chapter.action.download.add.button.selected')}
-                />
-                <SelectionFABActionItem
-                    action="delete"
-                    Icon={Delete}
-                    matchingItems={selectedChapters.filter(({ chapter }) => chapter.isDownloaded)}
-                    onClick={handleAction}
-                    title={t('chapter.action.download.delete.button.selected')}
-                />
-                <SelectionFABActionItem
-                    action="bookmark"
-                    Icon={BookmarkAdd}
-                    matchingItems={selectedChapters.filter(({ chapter }) => !chapter.isBookmarked)}
-                    onClick={handleAction}
-                    title={t('chapter.action.bookmark.add.button.selected')}
-                />
-                <SelectionFABActionItem
-                    action="unbookmark"
-                    Icon={BookmarkRemove}
-                    matchingItems={selectedChapters.filter(({ chapter }) => chapter.isBookmarked)}
-                    onClick={handleAction}
-                    title={t('chapter.action.bookmark.remove.button.selected')}
-                />
-                <SelectionFABActionItem
-                    action="mark_as_read"
-                    Icon={Done}
-                    matchingItems={selectedChapters.filter(({ chapter }) => !chapter.isRead)}
-                    onClick={handleAction}
-                    title={t('chapter.action.mark_as_read.add.button.selected')}
-                />
-                <SelectionFABActionItem
-                    action="mark_as_unread"
-                    Icon={RemoveDone}
-                    matchingItems={selectedChapters.filter(({ chapter }) => chapter.isRead)}
-                    onClick={handleAction}
-                    title={t('chapter.action.mark_as_read.remove.button.selected')}
-                />
+                {children(handleClose)}
             </Menu>
         </Box>
     );
